+++ src/konpy/views.clj
@@ -2,8 +2,7 @@
   (:require [hiccup2.core :as h]
             [ring.util.response :as response]))
 
-(def ^:private version "0.13.1")
-
+(def ^:private version "0.14.2")
 (defn base
   [content]
   [:html {:lang "en"}